requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
celery==5.6.2
redis==7.2.1
psycopg2-binary==2.9.9
ultralytics==8.2.18
torch==2.5.1
torchvision==0.20.1
python-multipart==0.0.9
aiofiles==23.2.1
